+++ CHANGES.rst
@@ -4,7 +4,9 @@ Histórico de Alterações
 1.5rc1 (unreleased)
 ^^^^^^^^^^^^^^^^^^^
 
-- Nothing changed yet.
+- Adiciona Products.RedirectionTool ao IDG;
+  habilita gerenciamento dos alias da redirection tool (plone.app.redirector).
+  [hvelarde]
 
 
 1.4 (2017-11-07)